requirements.txt
Django>=5.0,<5.1
django-htmx>=1.18.0
python-dotenv>=1.0
psycopg2-binary>=2.9
dj-database-url>=2.1.0
pytest>=8.2
pytest-django>=4.8
factory-boy>=3.3
ruff>=0.6
black>=24.8

